CORE-5114 move WireTx and its parts into common-internal for checkpointing this attempts
