--- libs/ledger/ledger-common-impl/src/main/java/net/corda/ledger/common/impl/transaction/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.ledger.common.impl.transaction;
-
-import org.osgi.annotation.bundle.Export;
