--- libs/ledger/ledger-common-impl/src/test/kotlin/net/corda/ledger/common/impl/transaction/serializer/WireTransactionKryoSerializerTest.kt
@@ -1,65 +1,0 @@
-package net.corda.ledger.common.impl.transaction.serializer
-
-import net.corda.application.impl.services.json.JsonMarshallingServiceImpl
-import net.corda.cipher.suite.impl.CipherSchemeMetadataImpl
-import net.corda.cipher.suite.impl.DigestServiceImpl
-import net.corda.crypto.merkle.impl.MerkleTreeFactoryImpl
-import net.corda.kryoserialization.testkit.createCheckpointSerializer
-import net.corda.ledger.common.impl.transaction.PrivacySaltImpl
-import net.corda.ledger.common.impl.transaction.WireTransaction
-import net.corda.ledger.common.testkit.WireTransactionExample.Companion.getWireTransaction
-import net.corda.v5.application.marshalling.JsonMarshallingService
-import net.corda.v5.cipher.suite.DigestService
-import net.corda.v5.crypto.merkle.MerkleTreeFactory
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.Test
-
-class WireTransactionKryoSerializerTest {
-    companion object {
-        private lateinit var digestService: DigestService
-        private lateinit var merkleTreeFactory: MerkleTreeFactory
-        private lateinit var jsonMarshallingService: JsonMarshallingService
-
-        @BeforeAll
-        @JvmStatic
-        fun setup() {
-            val schemeMetadata = CipherSchemeMetadataImpl()
-            digestService = DigestServiceImpl(schemeMetadata, null)
-            merkleTreeFactory = MerkleTreeFactoryImpl(digestService)
-            jsonMarshallingService = JsonMarshallingServiceImpl()
-        }
-    }
-
-    @Test
-    fun `serialization of a Wire Tx object using the kryo default serialization`() {
-        val wireTransaction = getWireTransaction(
-            digestService,
-            merkleTreeFactory,
-            jsonMarshallingService
-        )
-        val wireTransactionKryoSerializer = WireTransactionKryoSerializer(
-            merkleTreeFactory,
-            digestService,
-            jsonMarshallingService
-        )
-
-        val serializer = createCheckpointSerializer(
-            mapOf(
-                WireTransaction::class.java to wireTransactionKryoSerializer
-            ),
-            emptyList(),
-            setOf(PrivacySaltImpl::class.java)
-        )
-
-        val bytes = serializer.serialize(wireTransaction)
-        val deserialized = serializer.deserialize(bytes, WireTransaction::class.java)
-
-        assertThat(deserialized).isEqualTo(wireTransaction)
-        org.junit.jupiter.api.Assertions.assertDoesNotThrow {
-            deserialized.id
-        }
-        assertEquals(wireTransaction.id, deserialized.id)
-    }
-}
