+++ libs/ledger/ledger-common-internal/build.gradle
@@ -1,0 +1,33 @@
+plugins {
+    id 'corda.common-publishing'
+    id 'corda.osgi-test-conventions'
+    id 'corda.common-library'
+}
+
+description 'Corda ledger common - internal parts'
+
+dependencies {
+    compileOnly 'org.osgi:osgi.annotation'
+
+    implementation "net.corda.kotlin:kotlin-stdlib-jdk8-osgi"
+
+    implementation "net.corda:corda-ledger-common"
+    implementation platform("net.corda:corda-api:$cordaApiVersion")
+    implementation project(":libs:serialization:serialization-checkpoint-api")
+    implementation project(":libs:crypto:crypto-core")
+
+    testImplementation project(":testing:kryo-serialization-testkit")
+    testImplementation project(':testing:ledger:ledger-common-testkit')
+    testImplementation project(":libs:ledger:ledger-common-impl")
+    testImplementation project(":libs:serialization:serialization-kryo")
+    testImplementation project(":libs:crypto:merkle-impl")
+    testImplementation project(":libs:crypto:cipher-suite-impl")
+    testImplementation project(":libs:application:application-impl")
+
+    integrationTestImplementation project(':testing:sandboxes')
+    integrationTestImplementation project(":libs:serialization:serialization-checkpoint-api")
+
+    integrationTestRuntimeOnly project(':libs:lifecycle:lifecycle-impl')
+    integrationTestRuntimeOnly project(':libs:messaging:db-message-bus-impl')
+    integrationTestRuntimeOnly project(':libs:messaging:messaging-impl')
+}
