+++ libs/ledger/ledger-common-internal/test.bndrun
@@ -1,0 +1,15 @@
+-tester: biz.aQute.tester.junit-platform
+-runee: JavaSE-11
+-runfw: org.apache.felix.framework
+-resolve.effective: resolve,active
+-runtrace: true
+
+-runrequires: \
+    bnd.identity;id='net.corda.ledger-common-internal',\
+    bnd.identity;id='${project.archivesBaseName}-tests',\
+    bnd.identity;id='junit-jupiter-engine',\
+    bnd.identity;id='junit-platform-launcher',\
+    bnd.identity;id='net.corda.cipher-suite-impl',\
+    bnd.identity;id='net.corda.merkle-impl',\
+    bnd.identity;id='net.corda.serialization-kryo',\
+    bnd.identity;id='slf4j.simple'
