+++ settings.gradle
@@ -231,6 +231,7 @@ include 'libs:layered-property-map'
 include 'libs:layered-property-map:layered-property-map-test-converter'
 include 'libs:ledger:notary'
 include 'libs:ledger:ledger-common-impl'
+include 'libs:ledger:ledger-common-internal'
 include 'libs:ledger:ledger-common-serialization'
 include 'libs:ledger:ledger-consensual-impl'
 include 'libs:ledger:ledger-consensual-serialization'
